Put DTO Functions in Separate Class
