+++ vcell-core/src/main/java/cbit/vcell/field/io/FieldDataFileOperationSpec.java
@@ -10,21 +10,14 @@
 
 package cbit.vcell.field.io;
 
-import org.vcell.restclient.model.AnalyzedResultsFromFieldData;
+import cbit.vcell.math.VariableType;
+import cbit.vcell.solvers.CartesianMesh;
 import org.vcell.util.Extent;
 import org.vcell.util.ISize;
 import org.vcell.util.Origin;
 import org.vcell.util.document.ExternalDataIdentifier;
 import org.vcell.util.document.KeyValue;
 import org.vcell.util.document.User;
-
-import cbit.vcell.math.VariableType;
-import cbit.vcell.solvers.CartesianMesh;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
 
 /**
  * Insert the type's description here.
@@ -86,24 +79,6 @@ public FieldDataFileOperationSpec(short[][][] shortSpecData, double[][][] double
 	this.sourceSimParamScanJobIndex = sourceSimParamScanJobIndex;
 	this.sourceSimDataKey = sourceSimDataKey;
 	this.sourceOwner = sourceOwner;
-}
-
-public static AnalyzedResultsFromFieldData fieldDataSpecToAnalyzedResultsDTO(FieldDataFileOperationSpec fieldDataFileOperationSpec){
-	List<List<List<Integer>>> listVersion = Arrays.stream(fieldDataFileOperationSpec.shortSpecData) // Stream of short[][]
-			.map(twoDArray -> Arrays.stream(twoDArray) // Stream of short[]
-					.map(oneDArray -> {
-						List<Integer> list = new ArrayList<>();
-						for (short j : oneDArray) {
-							list.add((int) j);
-						}
-						return list;
-					}).collect(Collectors.toList())).collect(Collectors.toList());
-	AnalyzedResultsFromFieldData analyzedResultsFromFieldData = new AnalyzedResultsFromFieldData();
-	analyzedResultsFromFieldData.annotation(fieldDataFileOperationSpec.annotation); analyzedResultsFromFieldData.isize(ISize.iSizeToDTO(fieldDataFileOperationSpec.isize));
-	analyzedResultsFromFieldData.extent(Extent.extentToDTO(fieldDataFileOperationSpec.extent)); analyzedResultsFromFieldData.origin(Origin.originToDTO(fieldDataFileOperationSpec.origin));
-	analyzedResultsFromFieldData.times(Arrays.stream(fieldDataFileOperationSpec.times).boxed().toList()); analyzedResultsFromFieldData.setName(fieldDataFileOperationSpec.fieldDataName);
-	analyzedResultsFromFieldData.varNames(Arrays.stream(fieldDataFileOperationSpec.varNames).toList()); analyzedResultsFromFieldData.shortSpecData(listVersion);
-	return analyzedResultsFromFieldData;
 }
