+++ vcell-core/src/main/java/cbit/vcell/simdata/DataIdentifier.java
@@ -41,12 +41,6 @@ public DataIdentifier(String argName, VariableType argVariableType, Domain argDo
 	displayName = argDisplayName;
 }
 
-public static DataIdentifier dtoToDataIdentifier(org.vcell.restclient.model.DataIdentifier dto){
-	return new DataIdentifier(dto.getName(), VariableType.dtoToVariableType(dto.getVariableType()),
-			dto.getDomain() == null ? null : Domain.dtoToDomain(dto.getDomain()),
-			dto.getbFunction() == null ? false : dto.getbFunction(), dto.getDisplayName());
-}
-
 
 /**
  * Insert the method's description here.
