+++ vcell-core/src/main/java/org/vcell/util/document/ExternalDataIdentifier.java
@@ -26,22 +26,6 @@ public class ExternalDataIdentifier implements SimResampleInfoProvider,java.io.S
 	private org.vcell.util.document.User owner;
 	private String name;
 
-	public static org.vcell.restclient.model.ExternalDataIdentifier externalDataIdentifierToDTO(ExternalDataIdentifier externalDataIdentifier) {
-		org.vcell.restclient.model.ExternalDataIdentifier dto = new org.vcell.restclient.model.ExternalDataIdentifier();
-		dto.key(KeyValue.keyValueToDTO(externalDataIdentifier.getKey()));
-		dto.owner(User.userToDTO(externalDataIdentifier.getOwner()));
-		dto.name(externalDataIdentifier.getName());
-		return dto;
-	}
-
-	public static ExternalDataIdentifier dtoToExternalDataIdentifier(org.vcell.restclient.model.ExternalDataIdentifier dto){
-		ExternalDataIdentifier externalDataIdentifier = new ExternalDataIdentifier();
-		externalDataIdentifier.key = KeyValue.dtoToKeyValue(dto.getDataKey());
-		externalDataIdentifier.name = dto.getName();
-		externalDataIdentifier.owner = User.dtoToUser(dto.getOwner());
-		return externalDataIdentifier;
-	}
-
 /**
  * FieldDataIdentifier constructor comment.
  */
