+++ vcell-restclient/pom.xml
@@ -204,6 +204,11 @@
             <artifactId>vcell-util</artifactId>
             <version>0.0.1-SNAPSHOT</version>
         </dependency>
+        <dependency>
+            <groupId>org.vcell</groupId>
+            <artifactId>vcell-core</artifactId>
+            <version>0.0.1-SNAPSHOT</version>
+        </dependency>
 
         <!-- JSON processing: jackson -->
         <dependency>
