+++ vcell-util/pom.xml
@@ -138,7 +138,7 @@
 		<dependency>
 			<groupId>com.fasterxml.jackson.core</groupId>
 			<artifactId>jackson-annotations</artifactId>
-			<version>${jackson-version}</version>
+			<version>${jackson-annotations.version}</version>
 		</dependency>
 		<dependency>
 		  <groupId>com.sun.xml.bind</groupId>
